增加了持仓管理删除数量为0的持有

+++ law_test/src/warehouse/delet/deleteAction.java
@@ -3,8 +3,11 @@ package warehouse.delet;
 import org.json.JSONException;
 import org.json.JSONObject;
 import permission.manager.permissionChecker;
+import user.manager.User;
 import utils.dbOpener;
 import utils.sendManager;
+import utils.tokenChecker;
+import utils.tokenExtractor;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -27,6 +30,11 @@ public class deleteAction extends HttpServlet {
 
         if(!permissionChecker.checkPermissionAndResponse(request,response,this)) return;
         String stockid = request.getParameter("StockId");
+        String tocken= tokenExtractor.extractToken(request);
+        User user = tokenChecker.tokenToUser(tocken);
+        String a =null;
+        int quantity = Integer.parseInt(request.getParameter("Quantity"));
+        System.out.println("数量是："+quantity);
             System.out.println("前端页面传过来的ID是："+stockid+"<br>");
             request.setCharacterEncoding("UTF-8");
             try {
@@ -38,9 +46,17 @@ public class deleteAction extends HttpServlet {
                 Connection conn = dbOpener.getDB();
                 Statement statement = conn.createStatement();
                 System.out.println("Connect Database Ok！！！<br>");
-                String sql = "delete from tbl_userrealwh where StockId=" + stockid;
-                statement.executeUpdate(sql);
-                System.out.println(sql);
+                if(quantity==0)
+                {
+                    String sql = "delete from tbl_userrealwh where StockId= '"+stockid+"' and UserId='"+user.getUserId()+"'";
+                    statement.executeUpdate(sql);
+                    System.out.println(sql);
+                }
+                else
+                {
+                    sendManager.sendErrorJSONWithMsgAndCode(response,a,1);
+                    return;
+                }
                 statement.close();
                 conn.close();
                 System.out.println("Database Closed！！！<br>");
